requirements.txt
google-genai
pytest
python-dotenv
